Renames certificateFixture so that AVA will ignore.
